+++ spec/http_spec.rb
@@ -1,10 +1,14 @@
 require 'spec_helper'
 
-describe LeanplumApi::Connection::Production do
+describe LeanplumApi::Connection do
+  let(:http) { described_class.new(LeanplumApi.configuration.production_key) }
+  def argument_string(dev_mode)
+    "appId=#{LeanplumApi.configuration.app_id}&clientKey=#{LeanplumApi.configuration.production_key}&apiVersion=1.0.6&devMode=#{dev_mode}&action=multi&time=#{Time.now.utc.strftime('%s')}"
+  end
+
   context 'regular mode' do
     it 'should build the right multi url' do
-      http = described_class.new
-      expect(http.send(:authed_multi_param_string)).to eq("appId=#{LeanplumApi.configuration.app_id}&clientKey=#{LeanplumApi.configuration.production_key}&apiVersion=1.0.6&devMode=false&action=multi&time=#{Time.now.utc.strftime('%s')}")
+      expect(http.send(:authed_multi_param_string)).to eq(argument_string(false))
     end
   end
 
@@ -16,8 +20,7 @@ describe LeanplumApi::Connection::Production do
     end
 
     it 'should build the right developer mode url' do
-      http = described_class.new
-      expect(http.send(:authed_multi_param_string)).to eq("appId=#{LeanplumApi.configuration.app_id}&clientKey=#{LeanplumApi.configuration.production_key}&apiVersion=1.0.6&devMode=true&action=multi&time=#{Time.now.utc.strftime('%s')}")
+      expect(http.send(:authed_multi_param_string)).to eq(argument_string(true))
     end
   end
 end