+++ CHANGELOG.md
@@ -1,8 +1,11 @@
 ## 3.1.0
 Add the ability to send events as `userAttributes` properties
 
+## 3.0.3
+* Single connection class; rename `wait_for_job` to `wait_for_export_job`
+
 ## 3.0.2
-Leanplum changed their "Anomalous timestamp" message again, so now we're just going to reset everyone on any type of warning
+* Leanplum changed their "Anomalous timestamp" message again, so now we're just going to reset everyone on any type of warning
 
 ## 3.0.1
 * Remove `log_path` configuration option that should have come out in 3.0.0